Renaming the aux directory to avoid issues with Windows developers
