+++ src/main/scala/scalatutorial/sections/ImperativeProgramming.scala
@@ -5,7 +5,7 @@
 
 package scalatutorial.sections
 
-import scalatutorial.aux.BankAccount
+import scalatutorial.utils.BankAccount
 
 /** @param name imperative_programming */
 object ImperativeProgramming extends ScalaTutorialSection {
