+++ src/main/scala/scalatutorial/utils/Note.scala
@@ -3,6 +3,6 @@
  * Copyright (C) 2015-2016 47 Degrees, LLC. <http://www.47deg.com>
  */
 
-package scalatutorial.aux
+package scalatutorial.utils
 
 case class Note(name: String, duration: String, octave: Int)
