+++ src/main/scala/scalatutorial/utils/animals.scala
@@ -3,7 +3,7 @@
  * Copyright (C) 2015-2016 47 Degrees, LLC. <http://www.47deg.com>
  */
 
-package scalatutorial.aux
+package scalatutorial.utils
 
 trait Animal {
   def fitness: Int
